Refactor import paths and components in app/components directory
